+++ Changelog.md
@@ -14,6 +14,10 @@ Improved
 
 New
 
-- None
+- Added Authentication
+- Added Client for transfers
+- Added Client for contract agreements
+- Added Client for EDR Caches
+- Added Download Service
 
-# [v1.0.0] (main) (16/03/2026)
+# [v1.0.0] (main) (17/04/2026)